Fix setpoint reset bug (wip)
The problem seems to be that the update setpoints method was used
for just updating the "wrench setpoint" which is simply the control
output when running in open loop. However, the update setpoints metdhod
also integrated the position and orientation setpoints. That is, calling
update setpoints while holding the joystick to move the vehicle forward
would result in incrementing the position setpoint in the x-direction.
This is bad, because the control scheme we use is based on (for instance)
ignoring the x setpoint when trying to move in the x direction, and then
when we stop moving, it grabs the current state and sets it as the setpoint.

We should try running with this change in the simulator.