Update manipulator_interface to Otterholm's hardware
The manipulator_interface on otterholm will run on a nanopi neo 2+
with a stepper driver shield. Support for the raspberry pi and
beagelbone is therefore removed. The stepper driver is disabled
or enabled by toggeling the two PWM inputs to the stepper driver,
because the enable pin on the stepper driver is not accesible on the
stepper shield. For now there is only support for a single stepper
motor with direction control, the manipulator message is updated
accordingly. The joystick_interface is updated to the new manipulator
message.