MCU interface (#68)
* Add mcu_interface

This will add functionality to interface with the Giant Gecko.

SerialCom is a class that opens a device in /dev to interface with the MCU.
Callback functions are used to send and receive data when a ROS topic is
published to.

serial_rw.cpp sets up the ROS subscribers and calls ros::spin().

* Remove symbolic links, add correct files

* Move header files into include folder

* Add arming callback function

Create ROS subscriber 'arming_sub'. The callback function when publishing
to this topic is arming_callback. This sends the arming byte the same way
as the heartbeat message does.

* Correctly add arming function

To arm Manta, publish "ARM" to the /mcu_arm topic. Arming means setting
every PWM signal to 1500 for a finite amount of time (2-3 seconds).

Any other message than "ARM" will disarm the ROV, that is, setting all
PWM signals to 0.

* Map thruster PWM according to the setup in Manta

For now, create an "alternative_thruster_cmd[]" that maps the PWM signals
according to what thrusters are connected to the ESCs. This must be done
more elegantly in the future

* Add checksum in alternative command

* Correctly map thrusters

Note: this is to be done in the MCU in the future

* Remove thruster remapping

This is done in MCU instead.

* Introduce light_pwm_callback

Publishing to /mcu_light_pwm will set the corresponding PWM signal

* Change std:msgs from int32 to int16 in light callback

* Add mcu_interface node to launchfile. Publish heartbeat to MCU

* Flush serial input buffer. Read serial data in all callbacks

* Introduce is_leak() functionality

* Add uart read on leak message

* Add leak detection functionality

Also, read buffer size is increased so that larger messages can be read.

* Rename class to McuInterface

Also, edit CMakelists.txt

* Minor fixes

* Remove bottomside.launch from PR

* Remove cout streams. Change ROS_INFO to ROS_DEBUG

* Minor fix ROS_DEBUG vs ROS_INFO

For now, ROS_INFO is published at 1 Hz to notify users that we have
communicatin with MCU

* Add mail, licence

* Parameterize device name

* Add ROS_WARN where applicable.

Also, replace std::cout with ROS_ERROR

* Minor fix. Increase read buffer size